+++ lib/ui/router.dart
@@ -1,6 +1,4 @@
-
-import 'package:admin_web_app/ui/home/home_screen.dart';
-import 'package:admin_web_app/presentation/sample/my_home_page.dart';
+import 'package:admin_web_app/ui/sample/my_home_page.dart';
 import 'package:go_router/go_router.dart';
 
 final router = GoRouter(
